+++ Editor/CustomDictionary.cs
@@ -5,159 +5,34 @@ using UnityEngine;
 
 namespace DeltaField.Shaders.MeshHologram.Editor
 {
-    public class CustomDictionary
+    internal class CustomDictionary
     {
-        public Dictionary<CUSTOM_PROPERTY, CustomPropertyState> GetCustomProperties()
-        {
-            return new Dictionary<CUSTOM_PROPERTY, CustomPropertyState>
-            {
-                {CUSTOM_PROPERTY._Z_WRITE,new CustomPropertyState(
-                new List<string>(){"_"},
-                "_ZWrite", "Z Write")
-                },
-                {CUSTOM_PROPERTY._ACTIVATE_DIRECTIONAL_LIGHT_INFLUENCE,new CustomPropertyState(
-                new List<string>(){"_","_ACTIVATE_DIRECTIONALLIGHT_INFLUENCE"},
-                "_ActivateDirectionalLightInfluence", "Directional Light Influence")
-                },
-                {CUSTOM_PROPERTY._ACTIVATE_AMBIENT_INFLUENCE,new CustomPropertyState(
-                new List<string>(){"_","_ACTIVATE_AMBIENT_INFLUENCE"},
-                "_ActivateAmbientInfluence", "Ambient Influence")
-                },
-                {CUSTOM_PROPERTY._ACTIVATE_LIGHTVOLUMES_INFLUENCE,new CustomPropertyState(
-                new List<string>(){"_","_ACTIVATE_LIGHTVOLUMES"},
-                "_ActivateLightVolumesInfluence", "LightVolumes Influence")
-                },
-                {CUSTOM_PROPERTY._Z_CLIP,new CustomPropertyState(
-                new List<string>(){"_"},
-                "_ZClip", "Z Clip")
-                },
-                {CUSTOM_PROPERTY._ALPHA_TO_MASK,new CustomPropertyState(
-                new List<string>(){"_"},
-                "_AlphaToMaskMemory", "Alpha To Mask")
-                },
-                {CUSTOM_PROPERTY._USE_AUDIOLINK,new CustomPropertyState(
-                new List<string>(){"_","_USE_AUDIOLINK"},
-                "_UseAudioLink", "Use AudioLink")
-                },
-                {CUSTOM_PROPERTY._GEOMETRY_MESSY_TOGGLE,new CustomPropertyState(
-                new List<string>(){"_","_ACTIVATE_GEOMETRYMESSY"},
-                "_GeometryMessyToggle", "Activate Messy")
-                },
-
-                { CUSTOM_PROPERTY._FRAGMENT_SOURCE,new CustomPropertyState(
-                new List<string>(){
-                    "_FRAGMENTSOURCE_VALUE",
-                    "_FRAGMENTSOURCE_NOISE1ST",
-                    "_FRAGMENTSOURCE_NOISE2ND",
-                    "_FRAGMENTSOURCE_NOISE3RD",
-                    "_FRAGMENTSOURCE_AUDIOLINK_VU",
-                    "_FRAGMENTSOURCE_AUDIOLINK_CHRONOTENSITY",
-                },
-                "_FragmentSource", "Fragment Source")
-                },
-
-                { CUSTOM_PROPERTY._COLORING_SOURCE,new CustomPropertyState(
-                new List<string>(){
-                    "_COLORINGSOURCE_VALUE",
-                    "_COLORINGSOURCE_NOISE1ST",
-                    "_COLORINGSOURCE_NOISE2ND",
-                    "_COLORINGSOURCE_NOISE3RD",
-                    "_COLORINGSOURCE_AUDIOLINK_VU",
-                    "_COLORINGSOURCE_AUDIOLINK_CHRONOTENSITY",
-                },
-                "_ColoringSource", "Coloring Source")
-                },
-
-                { CUSTOM_PROPERTY._GEOMETRY_SOURCE,new CustomPropertyState(
-                new List<string>(){
-                    "_GEOMETRYSOURCE_VALUE",
-                    "_GEOMETRYSOURCE_NOISE1ST",
-                    "_GEOMETRYSOURCE_NOISE2ND",
-                    "_GEOMETRYSOURCE_NOISE3RD",
-                    "_GEOMETRYSOURCE_AUDIOLINK_VU",
-                    "_GEOMETRYSOURCE_AUDIOLINK_CHRONOTENSITY",
-                },
-                "_GeometrySource", "Geometry Source")
-                },
-
-                { CUSTOM_PROPERTY._GEOMETRY_MESSY_SOURCE,new CustomPropertyState(
-                new List<string>(){
-                    "_GEOMETRYMESSYSOURCE_VALUE",
-                    "_GEOMETRYMESSYSOURCE_NOISE1ST",
-                    "_GEOMETRYMESSYSOURCE_NOISE2ND",
-                    "_GEOMETRYMESSYSOURCE_NOISE3RD",
-                    "_GEOMETRYMESSYSOURCE_AUDIOLINK_VU",
-                    "_GEOMETRYMESSYSOURCE_AUDIOLINK_CHRONOTENSITY",
-                },
-                "_GeometryMessySource", "Messy Source")
-                },
-
-
-
-                { CUSTOM_PROPERTY._NOISE1ST_SPACE,new CustomPropertyState(
-                new List<string>(){
-                    "_NOISE1STSPACE_OFFSET",
-                    "_NOISE1STSPACE_MODEL",
-                    "_NOISE1STSPACE_WORLD",
-                    "_NOISE1STSPACE_ORIGIN_WORLD",
-                    "_NOISE1STSPACE_MODEL_WORLD",
-                    "_NOISE1STSPACE_VERTEX_COLOR",
-                    },
-                "_Noise1stSpace", "Space")
-                },
-                { CUSTOM_PROPERTY._NOISE2ND_SPACE,new CustomPropertyState(
-                new List<string>(){
-                    "_NOISE2NDSPACE_OFFSET",
-                    "_NOISE2NDSPACE_MODEL",
-                    "_NOISE2NDSPACE_WORLD",
-                    "_NOISE2NDSPACE_ORIGIN_WORLD",
-                    "_NOISE2NDSPACE_MODEL_WORLD",
-                    "_NOISE2NDSPACE_VERTEX_COLOR",
-                    },
-                "_Noise2ndSpace", "Space")
-                },
-                { CUSTOM_PROPERTY._NOISE3RD_SPACE,new CustomPropertyState(
-                new List<string>(){
-                    "_NOISE3RDSPACE_OFFSET",
-                    "_NOISE3RDSPACE_MODEL",
-                    "_NOISE3RDSPACE_WORLD",
-                    "_NOISE3RDSPACE_ORIGIN_WORLD",
-                    "_NOISE3RDSPACE_MODEL_WORLD",
-                    "_NOISE3RDSPACE_VERTEX_COLOR",
-                    },
-                "_Noise3rdSpace", "Space")
-                },
-
-                {CUSTOM_PROPERTY._RENDERING_MODE,new CustomPropertyState(
-                    new List<string>(){
-                        "_RENDERINGMODE_TRANSPARENT",
-                        "_RENDERINGMODE_CUTOUT",
-                        "_RENDERINGMODE_SHADOWONLY",
-                    },
-                    "_RenderingMode", "Rendering Mode"
-                )
-
-                }
-
-            };
-        }
-        public Dictionary<SHADER_PROPERTY, ShaderPropertyState> GetShaderProperties()
+        internal Dictionary<SHADER_PROPERTY, ShaderPropertyState> GetShaderProperties()
         {
             return new Dictionary<SHADER_PROPERTY, ShaderPropertyState>
             {
+                {SHADER_PROPERTY._RENDERING_MODE,new ShaderPropertyState(
+                    "_RenderingMode", "Rendering Mode")
+                },
                 {SHADER_PROPERTY._CULL,new ShaderPropertyState(
                     "_Cull", "Culling Mode")
                 },
-                {SHADER_PROPERTY._ZWRITE,new ShaderPropertyState(
+                {SHADER_PROPERTY._Z_WRITE,new ShaderPropertyState(
                     "_ZWrite", "Z Write")
                 },
-                {SHADER_PROPERTY._FORCED_Z_SCALE_ZERO,new ShaderPropertyState(
+                {SHADER_PROPERTY._CUSTOM_RENDER_QUEUE_T,new ShaderPropertyState(
+                    "_CustomRenderQueueT","Transparent Render Queue")
+                },
+                {SHADER_PROPERTY._CUSTOM_RENDER_QUEUE_C,new ShaderPropertyState(
+                    "_CustomRenderQueueC","Cutout Render Queue")
+                },
+                { SHADER_PROPERTY._FORCED_Z_SCALE_ZERO,new ShaderPropertyState(
                     "_Forced_Z_Scale_Zero", "Forced Z Scale Zero")
                 },
                 {SHADER_PROPERTY._BILLBOARD_MODE,new ShaderPropertyState(
                     "_BillboardMode", "Billboard Mode(Feature)")
                 },
-                { SHADER_PROPERTY._DISTANCE_INFLUENCE,new ShaderPropertyState(
+                {SHADER_PROPERTY._DISTANCE_INFLUENCE,new ShaderPropertyState(
                     "_DistanceInfluence", "Distance Influence")
                 },
                 {SHADER_PROPERTY._STEREO_MERGE_MODE,new ShaderPropertyState(
@@ -166,17 +41,53 @@ namespace DeltaField.Shaders.MeshHologram.Editor
                 {SHADER_PROPERTY._USE_FWIDTH,new ShaderPropertyState(
                     "_UseFwidth", "Use fwidth()")
                 },
+
+                {SHADER_PROPERTY._ACTIVATE_DIRECTIONAL_LIGHT_INFLUENCE, new ShaderPropertyState(
+                    "_ActivateDirectionalLightInfluence", "Directional Light Influence"
+                )},
                 {SHADER_PROPERTY._DIRECTIONAL_LIGHT_INFLUENCE,new ShaderPropertyState(
                     "_DirectionalLightInfluence", "")
                 },
+                {SHADER_PROPERTY._ACTIVATE_AMBIENT_INFLUENCE, new ShaderPropertyState(
+                    "_ActivateAmbientInfluence", "Ambient Influence"
+                )},
                 {SHADER_PROPERTY._AMBIENT_INFLUENCE,new ShaderPropertyState(
                     "_AmbientInfluence", "")
                 },
+                {SHADER_PROPERTY._ACTIVATE_LIGHTVOLUMES_INFLUENCE, new ShaderPropertyState(
+                    "_ActivateLightVolumesInfluence", "LightVolumes Influence"
+                )},
                 {SHADER_PROPERTY._LIGHTVOLUMES_INFLUENCE,new ShaderPropertyState(
                     "_LightVolumesInfluence", "")
                 },
                 {SHADER_PROPERTY._PREVIEW_MODE,new ShaderPropertyState(
                     "_PreviewMode", "Preview Mode")
+                },
+                {SHADER_PROPERTY._ANTI_ALIASING,new ShaderPropertyState(
+                    "_AntiAliasing", "Anti Aliasing")
+                },
+
+                {SHADER_PROPERTY._MAIN_TEX,new ShaderPropertyState(
+                    "_MainTex", "Fallback Texture")
+                },
+
+                {SHADER_PROPERTY._Z_CLIP,new ShaderPropertyState(
+                    "_ZClip","Z Clip")
+                },
+                {SHADER_PROPERTY._Z_TEST,new ShaderPropertyState(
+                    "_ZTest","Z Test")
+                },
+                {SHADER_PROPERTY._COLOR_MASK,new ShaderPropertyState(
+                    "_ColorMask","Color Mask")
+                },
+                {SHADER_PROPERTY._OFFSET_FACTOR,new ShaderPropertyState(
+                    "_OffsetFactor","Offset Factor")
+                },
+                {SHADER_PROPERTY._OFFSET_UNITS,new ShaderPropertyState(
+                    "_OffsetUnits","Offset Units")
+                },
+                {SHADER_PROPERTY._ALPHA_TO_MASK,new ShaderPropertyState(
+                    "_AlphaToMaskMemory", "Alpha To Mask")
                 },
 
                 {SHADER_PROPERTY._BLEND_OP,new ShaderPropertyState(
@@ -196,18 +107,6 @@ namespace DeltaField.Shaders.MeshHologram.Editor
                 },
                 {SHADER_PROPERTY._DST_BLEND_ALPHA,new ShaderPropertyState(
                     "_DstBlendAlpha", "Alpha Blend Mode Destination")
-                },
-                {SHADER_PROPERTY._Z_TEST,new ShaderPropertyState(
-                    "_ZTest","Z Test")
-                },
-                {SHADER_PROPERTY._COLOR_MASK,new ShaderPropertyState(
-                    "_ColorMask","Color Mask")
-                },
-                {SHADER_PROPERTY._OFFSET_FACTOR,new ShaderPropertyState(
-                    "_OffsetFactor","Offset Factor")
-                },
-                {SHADER_PROPERTY._OFFSET_UNITS,new ShaderPropertyState(
-                    "_OffsetUnits","Offset Units")
                 },
 
                 {SHADER_PROPERTY._STENCIL_REF,new ShaderPropertyState(
@@ -232,60 +131,11 @@ namespace DeltaField.Shaders.MeshHologram.Editor
                     "_StencilZFail","Stencil ZFail")
                 },
 
-                {SHADER_PROPERTY._MAIN_TEX,new ShaderPropertyState(
-                    "_MainTex", "Fallback Texture")
-                },
 
-                {SHADER_PROPERTY._AUDIOLINK_MASK_CONTROL_TEX,new ShaderPropertyState(
-                    "_AudioLinkMaskControlTex", "AudioLink Mask Control Tex")
+                {SHADER_PROPERTY._USE_AUDIOLINK,new ShaderPropertyState(
+                    "_UseAudioLink", "Use AudioLink")
                 },
-                {SHADER_PROPERTY._AUDIOLINK_MASK_CONTROL,new ShaderPropertyState(
-                    "_AudioLinkMaskControl", "AudioLink Mask Control Strength")
-                },
-                {SHADER_PROPERTY._FRAGMENT_MASK_CONTROL_TEX,new ShaderPropertyState(
-                    "_FragmentMaskControlTex", "Fragment Mask Control Tex")
-                },
-                {SHADER_PROPERTY._FRAGMENT_MASK_CONTROL,new ShaderPropertyState(
-                    "_FragmentMaskControl", "Fragment Mask Control Strength")
-                },
-                {SHADER_PROPERTY._COLORING_MASK_CONTROL_TEX,new ShaderPropertyState(
-                    "_ColoringMaskControlTex", "Coloring Mask Control Tex")
-                },
-                {SHADER_PROPERTY._COLORING_MASK_CONTROL,new ShaderPropertyState(
-                    "_ColoringMaskControl", "Coloring Mask Control Strength")
-                },
-                {SHADER_PROPERTY._GEOMETRY_MASK_CONTROL_TEX,new ShaderPropertyState(
-                    "_GeometryMaskControlTex", "Geometry Mask Control Tex")
-                },
-                {SHADER_PROPERTY._GEOMETRY_MASK_CONTROL,new ShaderPropertyState(
-                    "_GeometryMaskControl", "Geometry Mask Control Strength")
-                },
-                {SHADER_PROPERTY._GEOMETRY_MESSY_MASK_CONTROL_TEX,new ShaderPropertyState(
-                    "_GeometryMessyMaskControlTex", "Messy Mask Control Tex")
-                },
-                {SHADER_PROPERTY._GEOMETRY_MESSY_MASK_CONTROL,new ShaderPropertyState(
-                    "_GeometryMessyMaskControl", "Messy Mask Control Strength")
-                },
-                {SHADER_PROPERTY._NOISE1ST_OFFSET_CONTROL_TEX,new ShaderPropertyState(
-                    "_Noise1stMaskControlTex", "Noise 1st Offset Control Tex")
-                },
-                {SHADER_PROPERTY._NOISE1ST_OFFSET_CONTROL,new ShaderPropertyState(
-                    "_Noise1stMaskControl", "Noise 1st Offset Control Strength")
-                },
-                {SHADER_PROPERTY._NOISE2ND_OFFSET_CONTROL_TEX,new ShaderPropertyState(
-                    "_Noise2ndMaskControlTex", "Noise 2nd Offset Control Tex")
-                },
-                {SHADER_PROPERTY._NOISE2ND_OFFSET_CONTROL,new ShaderPropertyState(
-                    "_Noise2ndMaskControl", "Noise 2nd Offset Control")
-                },
-                {SHADER_PROPERTY._NOISE3RD_OFFSET_CONTROL_TEX,new ShaderPropertyState(
-                    "_Noise3rdMaskControlTex", "Noise 3rd Offset Control Tex")
-                },
-                {SHADER_PROPERTY._NOISE3RD_OFFSET_CONTROL,new ShaderPropertyState(
-                    "_Noise3rdMaskControl", "Noise 3rd Offset Control")
-                },
-
-                {SHADER_PROPERTY._AUDIOLINK_VU_BAND,new ShaderPropertyState(
+                { SHADER_PROPERTY._AUDIOLINK_VU_BAND,new ShaderPropertyState(
                     "_AudioLinkVUBand", "VU Band")
                 },
                 {SHADER_PROPERTY._AUDIOLINK_VU_SMOOTH,new ShaderPropertyState(
@@ -313,6 +163,7 @@ namespace DeltaField.Shaders.MeshHologram.Editor
                     "_AudioLinkThemeColorBand", "Theme Color Band")
                 },
 
+
                 {SHADER_PROPERTY._TRIANGLE_COMP,new ShaderPropertyState(
                     "_TriangleComp", "Triangle Compression")
                 },
@@ -325,6 +176,9 @@ namespace DeltaField.Shaders.MeshHologram.Editor
                 {SHADER_PROPERTY._LINE_GRADIENT_BIAS,new ShaderPropertyState(
                     "_LineGradientBias", "Line Gradient Bias")
                 },
+                {SHADER_PROPERTY._MANUAL_LINE_SCALING,new ShaderPropertyState(
+                    "_ManualLineScaling", "Manual Line Scaling")
+                },
                 {SHADER_PROPERTY._LINE_SCALE,new ShaderPropertyState(
                     "_LineScale", "Line Scale")
                 },
@@ -334,11 +188,15 @@ namespace DeltaField.Shaders.MeshHologram.Editor
                 {SHADER_PROPERTY._FRAGMENT_VALUE,new ShaderPropertyState(
                     "_FragmentValue", "Fragment Value")
                 },
+                {SHADER_PROPERTY._FRAGMENT_INVERSE,new ShaderPropertyState(
+                    "_FragmentInverse", "Fragment Inverse")
+                },
                 {SHADER_PROPERTY._PARTITION_TYPE,new ShaderPropertyState(
                     "_PartitionType", "Partition Type")
                 },
 
-                { SHADER_PROPERTY._COLOR0,new ShaderPropertyState(
+
+                {SHADER_PROPERTY._COLOR0,new ShaderPropertyState(
                     "_Color0", "Primary Color")
                 },
                 {SHADER_PROPERTY._COLOR1,new ShaderPropertyState(
@@ -360,59 +218,36 @@ namespace DeltaField.Shaders.MeshHologram.Editor
                     "_ColoringPartitionType", "Coloring Partition")
                 },
 
+
                 {SHADER_PROPERTY._GEOMETRY_VALUE,new ShaderPropertyState(
                     "_GeometryValue", "Geometry Value")
                 },
-                {SHADER_PROPERTY._GEOMETRY_PUSH_PULL_E,new ShaderPropertyState(
-                    "_GeometryPushPullE", "Activate Push/Pull")
+                {SHADER_PROPERTY._GEOMETRY_SCALE,new ShaderPropertyState(
+                    "_GeometryScale", "Activate Scale")
                 },
-                {SHADER_PROPERTY._GEOMETRY_PUSH_PULL,new ShaderPropertyState(
-                    "_GeometryPushPull", "Strength Push/Pull")
+                {SHADER_PROPERTY._GEOMETRY_SCALE_RANGE,new ShaderPropertyState(
+                    "_GeometryScaleRange", "Scale Range")
                 },
-                {SHADER_PROPERTY._GEOMETRY_PUSH_PULL_BIAS,new ShaderPropertyState(
-                    "_GeometryPushPullBias", "Bias Pull<=>Push")
+                {SHADER_PROPERTY._GEOMETRY_EXTRUDE,new ShaderPropertyState(
+                    "_GeometryExtrude", "Activate Extrude")
+                },
+                {SHADER_PROPERTY._GEOMETRY_EXTRUDE_RANGE,new ShaderPropertyState(
+                    "_GeometryExtrudeRange", "Extrude Range")
+                },
+                {SHADER_PROPERTY._GEOMETRY_ROTATION,new ShaderPropertyState(
+                    "_GeometryRotation", "Activate Rotation")
+                },
+                {SHADER_PROPERTY._GEOMETRY_ROTATION_INFLUENCE,new ShaderPropertyState(
+                    "_GeometryRotationInfluence", "Rotation Influence")
+                },
+                {SHADER_PROPERTY._GEOMETRY_ROTATION_REVERSE,new ShaderPropertyState(
+                    "_GeometryRotationReverse", "Rotation Reverse")
+                },
+                {SHADER_PROPERTY._GEOMETRY_ROTATION_NOISE_REPEAT,new ShaderPropertyState(
+                    "_GeometryRotationNoiseRepeat", "Rotation Noise Repeat")
                 },
 
-                {SHADER_PROPERTY._GEOMETRY_MESSY_VALUE,new ShaderPropertyState(
-                    "_GeometryMessyValue", "Messy Value")
-                },
-                {SHADER_PROPERTY._GEOMETRY_MESSY_SEED,new ShaderPropertyState(
-                    "_GeometryMessySeed", "Messy Primitive Seed")
-                },
-                {SHADER_PROPERTY._GEOMETRY_MESSY_ORBIT_ROTATION,new ShaderPropertyState(
-                    "_GeometryMessyOrbitRotation", "Messy Orbit Rotation")
-                },
-                {SHADER_PROPERTY._GEOMETRY_MESSY_ORBIT_ROTATION_FORWARD,new ShaderPropertyState(
-                    "_GeometryMessyOrbitRotationForward", "Messy Orbit Rotation Forward")
-                },
-                {SHADER_PROPERTY._GEOMETRY_MESSY_ORBIT_ROTATION_RIGHT,new ShaderPropertyState(
-                    "_GeometryMessyOrbitRotationRight", "Messy Orbit Rotation Right")
-                },
-                {SHADER_PROPERTY._GEOMETRY_MESSY_ORBIT_POSITION,new ShaderPropertyState(
-                    "_GeometryMessyOrbitPosition", "Messy Orbit Position")
-                },
-                {SHADER_PROPERTY._GEOMETRY_MESSY_ORBIT_SCALE_Y,new ShaderPropertyState(
-                    "_GeometryMessyOrbitScaleY", "Messy Orbit Scale Y")
-                },
-                {SHADER_PROPERTY._GEOMETRY_MESSY_ORBIT_SCALE_Z,new ShaderPropertyState(
-                    "_GeometryMessyOrbitScaleZ", "Messy Orbit Scale Z")
-                },
-                {SHADER_PROPERTY._GEOMETRY_MESSY_ORBIT_VARIANCE,new ShaderPropertyState(
-                    "_GeometryMessyOrbitVariance", "Messy Orbit Variance")
-                },
-                {SHADER_PROPERTY._GEOMETRY_MESSY_ORBIT_TIME_MULTIPLIER,new ShaderPropertyState(
-                    "_GeometryMessyTimeMultiplier", "Orbit Rotation Time Multiplier")
-                },
-                {SHADER_PROPERTY._GEOMETRY_MESSY_ORBIT_REFERENCE_TIME,new ShaderPropertyState(
-                    "_GeometryMessyReferenceTime", "Orbit Rotation Reference Time")
-                },
-                {SHADER_PROPERTY._ORBIT_ROTATION_REF_AUDIOLINK,new ShaderPropertyState(
-                    "_OrbitRotationRefAudioLink", "Orbit Rotation Reference AudioLink")
-                },
-                {SHADER_PROPERTY._GEOMETRY_MESSY_ORBIT_AUDIOLINK_STRENGTH,new ShaderPropertyState(
-                    "_GeometryMessyOrbitAudioLinkStrength", "Orbit Rotation AudioLink Strength")
-                },
-                { SHADER_PROPERTY._GEOMETRY_PARTITION_BIAS,new ShaderPropertyState(
+                {SHADER_PROPERTY._GEOMETRY_PARTITION_BIAS,new ShaderPropertyState(
                     "_GeometryPartitionBias", "Geometry Partition Bias| Vertex <=> Center")
                 },
                 {SHADER_PROPERTY._PIXELIZATION_SPACE,new ShaderPropertyState(
@@ -422,7 +257,207 @@ namespace DeltaField.Shaders.MeshHologram.Editor
                     "_Pixelization", "Vertex Pixelization")
                 },
 
-                { SHADER_PROPERTY._NOISE1ST_OFFSET0,new ShaderPropertyState(
+
+                {SHADER_PROPERTY._ACTIVATE_ORBIT,new ShaderPropertyState(
+                    "_ActivateOrbit", "Activate Orbit")
+                },
+
+                {SHADER_PROPERTY._ORBIT_VALUE,new ShaderPropertyState(
+                    "_OrbitValue", "Orbit Value")
+                },
+                {SHADER_PROPERTY._ORBIT_SEED,new ShaderPropertyState(
+                    "_OrbitSeed", "Orbit Seed")
+                },
+                {SHADER_PROPERTY._ORBIT_PRIMITIVE_THRESHOLD,new ShaderPropertyState(
+                    "_OrbitPrimitiveThreshold", "Orbit Primitive Threshold")
+                },
+                {SHADER_PROPERTY._ORBIT_ROTATION_VALUE,new ShaderPropertyState(
+                    "_OrbitRotationValue", "Orbit Rotation Value")
+                },
+                {SHADER_PROPERTY._ORBIT_ROTATION_SEED,new ShaderPropertyState(
+                    "_OrbitRotationSeed", "Orbit Rotation Seed")
+                },
+                {SHADER_PROPERTY._ORBIT_ROTATION,new ShaderPropertyState(
+                    "_OrbitRotation", "Orbit Rotation")
+                },
+                {SHADER_PROPERTY._ORBIT_ROTATION_TIME_MULTIPLIER,new ShaderPropertyState(
+                    "_OrbitRotationTimeMultiplier", "Orbit Rotation Time Multiplier")
+                },
+                {SHADER_PROPERTY._ORBIT_ROTATION_VARIANCE,new ShaderPropertyState(
+                    "_OrbitRotationVariance", "Orbit Rotation Variance")
+                },
+
+                {SHADER_PROPERTY._ORBIT_ROTATION_REF_AUDIOLINK,new ShaderPropertyState(
+                    "_OrbitRotationRefAudioLink", "Orbit Rotation Reference AudioLink")
+                },
+                {SHADER_PROPERTY._ORBIT_ROTATION_AUDIOLINK_STRENGTH,new ShaderPropertyState(
+                    "_OrbitRotationAudioLinkStrength", "Orbit Rotation AudioLink Strength")
+                },
+
+                {SHADER_PROPERTY._ORBIT_OFFSET,new ShaderPropertyState(
+                    "_OrbitOffset", "Orbit Offset")
+                },
+                {SHADER_PROPERTY._ORBIT_SCALE,new ShaderPropertyState(
+                    "_OrbitScale", "Orbit Scale")
+                },
+
+                {SHADER_PROPERTY._ORBIT_WAVE_STRENGTH,new ShaderPropertyState(
+                    "_OrbitWaveStrength", "Orbit Wave Strength")
+                },
+                {SHADER_PROPERTY._ORBIT_WAVE_FREQUENCY,new ShaderPropertyState(
+                    "_OrbitWaveFrequency", "Orbit Wave Frequency")
+                },
+                {SHADER_PROPERTY._ORBIT_WAVE_PHASE,new ShaderPropertyState(
+                    "_OrbitWavePhase", "Orbit Wave Phase")
+                },
+                {SHADER_PROPERTY._ORBIT_WAVE_TIME_MULTIPLIER,new ShaderPropertyState(
+                    "_OrbitWaveTimeMultiplier", "Orbit Wave Time Multiplier")
+                },
+
+                {SHADER_PROPERTY._ORBIT_WAVE_REF_AUDIOLINK,new ShaderPropertyState(
+                    "_OrbitWaveRefAudioLink", "Orbit Wave Reference AudioLink")
+                },
+                {SHADER_PROPERTY._ORBIT_WAVE_AUDIOLINK_STRENGTH,new ShaderPropertyState(
+                    "_OrbitWaveAudioLinkStrength", "Orbit Wave AudioLink Strength")
+                },
+                {SHADER_PROPERTY._ORBIT_WAVE_AUDIOLINK_SPECTRUM_MIRROR,new ShaderPropertyState(
+                    "_OrbitWaveAudioLinkSpectrumMirror", "AudioLink Spectrum Mirror")
+                },
+                {SHADER_PROPERTY._ORBIT_WAVE_AUDIOLINK_SPECTRUM_TYPE,new ShaderPropertyState(
+                    "_OrbitWaveAudioLinkSpectrumType", "AudioLink Spectrum Type")
+                },
+                {SHADER_PROPERTY._ORBIT_WAVE_AUDIOLINK_SPECTRUM_RANGE,new ShaderPropertyState(
+                    "_OrbitWaveAudioLinkSpectrumRange", "AudioLink Spectrum Range")
+                },
+
+                {SHADER_PROPERTY._FRAGMENT_SOURCE,new ShaderPropertyState(
+                    "_FragmentSource", "Fragment Source")
+                },
+                {SHADER_PROPERTY._COLORING_SOURCE,new ShaderPropertyState(
+                    "_ColoringSource", "Coloring Source")
+                },
+                {SHADER_PROPERTY._GEOMETRY_SOURCE,new ShaderPropertyState(
+                    "_GeometrySource", "Geometry Source")
+                },
+                {SHADER_PROPERTY._ORBIT_SOURCE,new ShaderPropertyState(
+                    "_OrbitSource", "Orbit Source")
+                },
+                {SHADER_PROPERTY._ORBIT_ROTATION_SOURCE,new ShaderPropertyState(
+                    "_OrbitRotationSource", "Orbit Rotation Source")
+                },
+
+
+                {SHADER_PROPERTY._AUDIOLINK_MASK_CONTROL_TEX,new ShaderPropertyState(
+                    "_AudioLinkMaskControlTex", "AudioLink Mask Control Tex")
+                },
+                {SHADER_PROPERTY._AUDIOLINK_MASK_CONTROL,new ShaderPropertyState(
+                    "_AudioLinkMaskControl", "AudioLink Mask Control Strength")
+                },
+                {SHADER_PROPERTY._FRAGMENT_MASK_CONTROL_TEX,new ShaderPropertyState(
+                    "_FragmentMaskControlTex", "Fragment Mask Control Tex")
+                },
+                {SHADER_PROPERTY._FRAGMENT_MASK_CONTROL,new ShaderPropertyState(
+                    "_FragmentMaskControl", "Fragment Mask Control Strength")
+                },
+                {SHADER_PROPERTY._COLORING_MASK_CONTROL_TEX,new ShaderPropertyState(
+                    "_ColoringMaskControlTex", "Coloring Mask Control Tex")
+                },
+                {SHADER_PROPERTY._COLORING_MASK_CONTROL,new ShaderPropertyState(
+                    "_ColoringMaskControl", "Coloring Mask Control Strength")
+                },
+                {SHADER_PROPERTY._GEOMETRY_MASK_CONTROL_TEX,new ShaderPropertyState(
+                    "_GeometryMaskControlTex", "Geometry Mask Control Tex")
+                },
+                {SHADER_PROPERTY._GEOMETRY_MASK_CONTROL,new ShaderPropertyState(
+                    "_GeometryMaskControl", "Geometry Mask Control Strength")
+                },
+                {SHADER_PROPERTY._ORBIT_MASK_CONTROL_TEX,new ShaderPropertyState(
+                    "_OrbitMaskControlTex", "Orbit Control Tex")
+                },
+                {SHADER_PROPERTY._ORBIT_MASK_CONTROL,new ShaderPropertyState(
+                    "_OrbitMaskControl", "Orbit Control Strength")
+                },
+                {SHADER_PROPERTY._NOISE1ST_OFFSET_CONTROL_TEX,new ShaderPropertyState(
+                    "_Noise1stOffsetControlTex", "Noise 1st Offset Control Tex")
+                },
+                {SHADER_PROPERTY._NOISE1ST_OFFSET_CONTROL,new ShaderPropertyState(
+                    "_Noise1stOffsetControl", "Noise 1st Offset Control Strength")
+                },
+                {SHADER_PROPERTY._NOISE2ND_OFFSET_CONTROL_TEX,new ShaderPropertyState(
+                    "_Noise2ndOffsetControlTex", "Noise 2nd Offset Control Tex")
+                },
+                {SHADER_PROPERTY._NOISE2ND_OFFSET_CONTROL,new ShaderPropertyState(
+                    "_Noise2ndOffsetControl", "Noise 2nd Offset Control")
+                },
+                {SHADER_PROPERTY._NOISE3RD_OFFSET_CONTROL_TEX,new ShaderPropertyState(
+                    "_Noise3rdOffsetControlTex", "Noise 3rd Offset Control Tex")
+                },
+                {SHADER_PROPERTY._NOISE3RD_OFFSET_CONTROL,new ShaderPropertyState(
+                    "_Noise3rdOffsetControl", "Noise 3rd Offset Control")
+                },
+
+
+                {SHADER_PROPERTY._FRAGMENT_AUDIOLINK_NOISE_SPECTRUM,new ShaderPropertyState(
+                    "_FragmentAudioLinkNoiseSpectrum", "Fragment AudioLink Noise Spectrum")
+                },
+                {SHADER_PROPERTY._COLORING_AUDIOLINK_NOISE_SPECTRUM,new ShaderPropertyState(
+                    "_ColoringAudioLinkNoiseSpectrum", "Coloring AudioLink Noise Spectrum")
+                },
+                {SHADER_PROPERTY._GEOMETRY_AUDIOLINK_NOISE_SPECTRUM,new ShaderPropertyState(
+                    "_GeometryAudioLinkNoiseSpectrum", "Geometry AudioLink Noise Spectrum")
+                },
+
+                {SHADER_PROPERTY._FRAGMENT_AUDIOLINK_STRENGTH,new ShaderPropertyState(
+                    "_FragmentAudioLinkStrength", "Noise Spectrum Strength")
+                },
+                {SHADER_PROPERTY._FRAGMENT_AUDIOLINK_SPECTRUM_MIRROR,new ShaderPropertyState(
+                    "_FragmentAudioLinkSpectrumMirror", "Noise Spectrum Mirror")
+                },
+                {SHADER_PROPERTY._FRAGMENT_AUDIOLINK_SPECTRUM_TYPE,new ShaderPropertyState(
+                    "_FragmentAudioLinkSpectrumType", "Noise Spectrum Type")
+                },
+                {SHADER_PROPERTY._FRAGMENT_AUDIOLINK_SPECTRUM_RANGE,new ShaderPropertyState(
+                    "_FragmentAudioLinkSpectrumRange", "Noise Spectrum Range")
+                },
+
+                {SHADER_PROPERTY._COLORING_AUDIOLINK_STRENGTH,new ShaderPropertyState(
+                    "_ColoringAudioLinkStrength", "Noise Spectrum Strength")
+                },
+                {SHADER_PROPERTY._COLORING_AUDIOLINK_SPECTRUM_MIRROR,new ShaderPropertyState(
+                    "_ColoringAudioLinkSpectrumMirror", "Noise Spectrum Mirror")
+                },
+                {SHADER_PROPERTY._COLORING_AUDIOLINK_SPECTRUM_TYPE,new ShaderPropertyState(
+                    "_ColoringAudioLinkSpectrumType", "Noise Spectrum Type")
+                },
+                {SHADER_PROPERTY._COLORING_AUDIOLINK_SPECTRUM_RANGE,new ShaderPropertyState(
+                    "_ColoringAudioLinkSpectrumRange", "Noise Spectrum Range")
+                },
+
+                {SHADER_PROPERTY._GEOMETRY_AUDIOLINK_STRENGTH,new ShaderPropertyState(
+                    "_GeometryAudioLinkStrength", "Noise Spectrum Strength")
+                },
+                {SHADER_PROPERTY._GEOMETRY_AUDIOLINK_SPECTRUM_MIRROR,new ShaderPropertyState(
+                    "_GeometryAudioLinkSpectrumMirror", "Noise Spectrum Mirror")
+                },
+                {SHADER_PROPERTY._GEOMETRY_AUDIOLINK_SPECTRUM_TYPE,new ShaderPropertyState(
+                    "_GeometryAudioLinkSpectrumType", "Noise Spectrume Type")
+                },
+                {SHADER_PROPERTY._GEOMETRY_AUDIOLINK_SPECTRUM_RANGE,new ShaderPropertyState(
+                    "_GeometryAudioLinkSpectrumRange", "Noise Spectrum Range")
+                },
+
+
+                {SHADER_PROPERTY._NOISE1ST_SPACE,new ShaderPropertyState(
+                    "_Noise1stSpace", "Space")
+                },
+                {SHADER_PROPERTY._NOISE2ND_SPACE,new ShaderPropertyState(
+                    "_Noise2ndSpace", "Space")
+                },
+                {SHADER_PROPERTY._NOISE3RD_SPACE,new ShaderPropertyState(
+                    "_Noise3rdSpace", "Space")
+                },
+
+                {SHADER_PROPERTY._NOISE1ST_OFFSET0,new ShaderPropertyState(
                     "_Noise1stOffset0", "Offset")
                 },
                 {SHADER_PROPERTY._NOISE1ST_SCALE0,new ShaderPropertyState(
@@ -460,9 +495,6 @@ namespace DeltaField.Shaders.MeshHologram.Editor
                 },
                 {SHADER_PROPERTY._NOISE1ST_PHASE_SCALE,new ShaderPropertyState(
                     "_Noise1stPhaseScale", "Phase Scale")
-                },
-                {SHADER_PROPERTY._NOISE1ST_REFERENCE_TIME,new ShaderPropertyState(
-                    "_Noise1stReferenceTime", "Reference Time")
                 },
                 {SHADER_PROPERTY._NOISE1ST_PHASE_REF_AUDIOLINK,new ShaderPropertyState(
                     "_Noise1stPhaseRefAudioLink", "Phase Reference AudioLink")
@@ -507,9 +539,6 @@ namespace DeltaField.Shaders.MeshHologram.Editor
                 {SHADER_PROPERTY._NOISE2ND_PHASE_SCALE,new ShaderPropertyState(
                     "_Noise2ndPhaseScale", "Phase Scale")
                 },
-                {SHADER_PROPERTY._NOISE2ND_REFERENCE_TIME,new ShaderPropertyState(
-                    "_Noise2ndReferenceTime", "Reference Time")
-                },
                 {SHADER_PROPERTY._NOISE2ND_PHASE_REF_AUDIOLINK,new ShaderPropertyState(
                     "_Noise2ndPhaseRefAudioLink", "Phase Reference AudioLink")
                 },
@@ -553,23 +582,12 @@ namespace DeltaField.Shaders.MeshHologram.Editor
                 {SHADER_PROPERTY._NOISE3RD_PHASE_SCALE,new ShaderPropertyState(
                     "_Noise3rdPhaseScale", "Phase Scale")
                 },
-                {SHADER_PROPERTY._NOISE3RD_REFERENCE_TIME,new ShaderPropertyState(
-                    "_Noise3rdReferenceTime", "Reference Time")
-                },
                 {SHADER_PROPERTY._NOISE3RD_PHASE_REF_AUDIOLINK,new ShaderPropertyState(
                     "_Noise3rdPhaseRefAudioLink", "Phase Reference AudioLink")
-                },
-
-                { SHADER_PROPERTY._CUSTOM_RENDER_QUEUE_T,new ShaderPropertyState(
-                    "_CustomRenderQueueT","Transparent Render Queue")
-                },
-                {SHADER_PROPERTY._CUSTOM_RENDER_QUEUE_C,new ShaderPropertyState(
-                    "_CustomRenderQueueC","Cutout Render Queue")
                 }
-
             };
         }
-        public static Dictionary<CUSTOM_GUI, GUIStyle> gui = new Dictionary<CUSTOM_GUI, GUIStyle> {
+        internal static Dictionary<CUSTOM_GUI, GUIStyle> gui = new Dictionary<CUSTOM_GUI, GUIStyle> {
             {CUSTOM_GUI.PLAIN,new GUIStyle(){
 
             }},
@@ -580,25 +598,15 @@ namespace DeltaField.Shaders.MeshHologram.Editor
                 richText = true, alignment = TextAnchor.UpperCenter, fontSize = 24
             }},
             {CUSTOM_GUI.HEADER0,new GUIStyle(){
-                richText = true, alignment = TextAnchor.UpperLeft, fontSize = 16
+                richText = true, alignment = TextAnchor.UpperLeft, fontSize = 16, fontStyle=FontStyle.Bold
+            }},
+            {CUSTOM_GUI.HEADER1,new GUIStyle(){
+                richText = true, alignment = TextAnchor.UpperLeft, fontSize = 12, fontStyle=FontStyle.Bold
             }}
         };
-
     }
-    public class CustomPropertyState {
-        public MaterialProperty var;
-        readonly public string property;
-        public string display;
-        readonly public List<string> keywords;
-        public int value = 0;
-        public CustomPropertyState(List<string> k, string p, string d)
-        {
-            keywords = k;
-            property = p;
-            display = d;
-        }
-    }
-    public class ShaderPropertyState {
+    internal class ShaderPropertyState
+    {
         public MaterialProperty var;
         readonly public string property;
         public string display;